Get Pipes working again.
They stopped working because of the change to container node IDs (and rendered nodes IDs for containers).  The UI was comparing the IDs, which was never safe.  I have just removed that code.  This does leave the possibility of us having the pipe control operation take a long time, the user navigate to a different node, and then the terminal pop up, but I think thats better that teaching the UI to understand the format of the node IDs.