Separate node colors
Node colors are procedurally generated by a node's rank. This can be a
string like `weaveworks/lb-1`. Now if all the nodes share the same
prefix in their rank, e.g. `weaveworks/` they will all get the same
color.

An approach to separate colors a bit more would be to remove the longest
common rank prefix and set that as the color key.

Pros:

- colors are better separation for homogeneous rank values

Cons:

- if 99 nodes share a prefix and one does not, the 99 will be colored
  similarly (which is what happens now)
- if that one node appears and dissappears, the 99 nodes' colors will
  switch based on their new LCP
- Loading details panels from other topologies may yield a different
  color because that node is loaded in isolation (LCP of that
topology's nodes is undetermined)