in target.Merge(arg), always merge arg into target
That way, as the top-level target.Merge(arg) on Report recursively
descends into the data structure, all component Merge operations are
invoked as target_component.Merge(arg_component).

Merge is commutative, so semantically this does not matter. But it is
easier to reason about and debug Merge code this way.

Moreover, the merge order matters for performance since some merges, in
particular LatestMap.Merge, are faster when the entries in the target
are younger than the entries in the arg. So this change allows
high-level code to invoke Report.Merge in the optimal order, in the
knowledge that the order is preserved for component merges.

Most of the Merge functions were already preserving order. Of the four
that didn't, Nodes.Merge is the most significant - it is the only one
that affects LatestMap.Merge order.