controls: plugins can add generic controls to start a container
This patch adds a way for plugins to add controls to start a new
container such as GDB for debugging. This does not require the plugin to
know the list of Docker containers, nor does it require to communicate
with the Docker daemon at all. The plumbing is done in the probe itself.
The plugin merely advertises the name of the GDB-like container to use.

This is achieved through two new fields in the report that the plugin
sends to the probe.

```
    "controls": {
      "ctrl-one": {
        "id": "ctrl-one",
        "human": "Ctrl One",
        "icon": "fa-futbol-o",
        "rank": 1,
        "always_propagated": true,
        "start_image": "albanc/toolbox"
      },
    }
```

Notice the two new optional fields:

- AlwaysPropagated (default: false): this asks the new control tagger
(probe/controls/tagger.go) to add the control on all nodes of the topology. In
this way, the plugin does not need to know the list of nodes and it does not
need to duplicate the work from the probe.

- StartImage (default: empty): if not empty, the probe will not forward the
command to the plugin but do the job itself instead. It creates a new Docker
container from the specified image and starts it. A window in the Scope UI is
immediately created and attached to the container.

The plugin developer can simply create an image that will start a
debugger such as GDB or strace. The container will be created with the
following environment variables: $SCOPE_PLUGIN_ID, $SCOPE_CONTROL_ID,
$SCOPE_IMAGE, $SCOPE_APP_ID, $SCOPE_NODE_ID. From $SCOPE_NODE_ID, the
image can determine which process to gdb or strace.

Previously, the controlID given by the plugin was converted before being
added in the report in the following form: "pluginID~controlID". Now, if
the StartImage field is not empty, the controlID added in the report
will now have this form: "pluginID~controlID~encodedImage". encodedImage
uses url.QueryEscape() to make sure that slashes in image names don't
bother Scope App.