tools/sched: Fix unicode encoding issues
The response.text and json-parsed string values are both unicode strings.
The print statement requires bytes, as per write(2).
Python is dumb and bad at converting between them properly - in the name of being "safe"
it errors if any non-ascii characters are present, so it works until it doesn't.
We fix this by explicitly encoding as utf-8.