report: encode timestamps in a compact form
This patch changes the encoding of two kinds of timestamps:
- timestamps in the "latest" map
- timestamps in metrics

We use Golang's internal time.Time marshalling and then base64 on it.

Before this patch, this is how it looked:
```json
"host_node_id" : {
   "value" : "neptune;<host>",
   "timestamp" : "2017-04-07T15:50:55.320041355Z"
},
...
"samples" : [
   {
      "date" : "2017-04-07T17:50:40.284167472+02:00",
      "value" : 249266176
   },
```

After this patch, this is how it looks:
```json
"latest" : {
   "host_node_id" : {
      "t" : "AQAAAA7QgZARDlWbvv//",
      "value" : "neptune;<host>"
   },
...
"samples" : [
   {
      "v" : 24.330900243309,
      "t" : "AQAAAA7QgZANEJKTv///"
   },
   {
      "v" : 25.9842519685039,
      "t" : "AQAAAA7QgZAODml7Af//"
   },
```

On my laptop with one Docker container running and 392 processes
running, I see 17815 timestamps in the report.

This should make the report smaller in its uncompressed form.

However, I should check if it makes the reports bigger in its compressed
form. It might happen because base64 strings are more difficult to
compress. I have not tested this yet. See:
http://stackoverflow.com/questions/38124361/why-does-base64-encoded-data-compress-so-bad